Create sudoku_grid
Description: This Python file defines the initial Sudoku grid to be solved. It represents the puzzle as a 2D list where empty cells are represented as 0s.
Usage: You can use this file to change the initial Sudoku puzzle by modifying the grid variable with your own puzzle. The solver (in algorithm.py) will attempt to solve the puzzle based on this input.